requirements.txt
streamlit
python-dotenv
requests
langchain-core
langchain-google-genai
